python/requirements.txt
pyserial==3.5
matplotlib>=3.8
numpy>=1.26

